fixed registration issue
